fix : 파일구조 정리(manage)
