+++ src/app/manage/manageDocument/modal/Footer.tsx
