+++ src/app/manage/managePassOrFail/page.tsx
@@ -1,0 +1,12 @@
+"use client";
+
+// 합불관리 화면
+// 담당자(담당 브랜치): yerim
+
+interface ManagePassOrFailProps {
+  onPrev: () => void;
+}
+
+export default function ManagePassOrFail({ onPrev }: ManagePassOrFailProps) {
+  return <div>합불관리</div>;
+}
