--- src/app/manageDocument/MozipManageBlocks/MozipManageInterview.tsx
@@ -1,11 +1,0 @@
-interface ManageInterviewProps {
-  onNext: () => void;
-  onPrev: () => void;
-}
-
-export default function MozipManageInterview({
-  onNext,
-  onPrev,
-}: ManageInterviewProps) {
-  return <div>면접 관리</div>;
-}
