--- src/app/manageDocument/MozipManageBlocks/MozipManagePassOrFail.tsx
@@ -1,9 +1,0 @@
-interface ManagePassOrFailProps {
-  onPrev: () => void;
-}
-
-export default function MozipManagePassOrFail({
-  onPrev,
-}: ManagePassOrFailProps) {
-  return <div>합불관리</div>;
-}
