--- src/app/manageInterview/page.tsx
@@ -1,10 +1,0 @@
-// 면접관리 화면
-// 담당자(담당 브랜치): yerim
-
-export default function Home() {
-	return (
-		<div>
-			면접관리 화면
-		</div>
-	);
-}
